+++ core/src/main/kotlin/behavior/interaction/response/InteractionResponseBehavior.kt
@@ -1,15 +1,14 @@
-package dev.kord.core.behavior.interaction
+package dev.kord.core.behavior.interaction.response
 
 import dev.kord.common.entity.Snowflake
 import dev.kord.core.KordObject
 import dev.kord.core.cache.data.toData
 import dev.kord.core.entity.Message
 import dev.kord.core.entity.Strategizable
-import dev.kord.core.entity.interaction.EphemeralFollowupMessage
-import dev.kord.core.entity.interaction.PublicFollowupMessage
+import dev.kord.core.entity.interaction.followup.EphemeralFollowupMessage
+import dev.kord.core.entity.interaction.followup.PublicFollowupMessage
 import dev.kord.core.exception.EntityNotFoundException
 import dev.kord.rest.builder.message.create.FollowupMessageCreateBuilder
-import dev.kord.rest.builder.message.modify.InteractionResponseModifyBuilder
 import dev.kord.rest.request.RestRequestException
 import kotlin.contracts.InvocationKind
 import kotlin.contracts.contract
@@ -17,7 +16,7 @@ import kotlin.contracts.contract
 /**
  * The behavior of a [Discord ActionInteraction Response](https://discord.com/developers/docs/interactions/receiving-and-responding#responding-to-an-interaction)
  */
-public interface InteractionResponseBehavior : KordObject, Strategizable {
+public sealed interface InteractionResponseBehavior : KordObject, Strategizable {
     public val applicationId: Snowflake
     public val token: String
 
@@ -56,19 +55,4 @@ public suspend inline fun InteractionResponseBehavior.followUpEphemeral(builder:
     contract { callsInPlace(builder, InvocationKind.EXACTLY_ONCE) }
     val message = kord.rest.interaction.createFollowupMessage(applicationId, token, ephemeral = true, builder)
     return EphemeralFollowupMessage(Message(message.toData(), kord), applicationId, token, kord)
-}
-
-/**
- * Requests to edit this interaction response.
- *
- * @return The edited [Message] of the interaction response.
- *
- * @throws RestRequestException if something went wrong during the request.
- */
-public suspend inline fun InteractionResponseBehavior.edit(
-    builder: InteractionResponseModifyBuilder.() -> Unit,
-): Message {
-    contract { callsInPlace(builder, InvocationKind.EXACTLY_ONCE) }
-    val message = kord.rest.interaction.modifyInteractionResponse(applicationId, token, builder)
-    return Message(message.toData(), kord)
 }